Add in a link to a Code of Conduct (#244)
* Add in a link to a Code of Conduct

If the OIF Code of Conduct is the wrong one, what is the correct one?

This should be easily findable in the footer or a main menu, depending on information architecture of the site.

I am proposing this link in this location as my best guest of what CoC to use and where to put the link.

If there should be a different CoC or location for the link, feel free to use this merge request to and patch it for the correct info.

* Fixing changing link text

I misread the markup and didn't notice my original copy/paste included one more value to change!