Updates 42e82eb978b873def16c96be3c6d3fb46e63bd4b: Canonical URL for our CoC (#290)
* Canonical URL for our CoC

Although the community is using a code of conduct located on another
domain, this lets us have a forever-URL we can always use, then either
use as a land-and-bounce page, or replace with any future CoC we
create.

Signed-off-by: Karsten Wade <kwade@redhat.com>

* Better link this way

This should make https://operate-first.cloud/code_of_conduct work.

Signed-off-by: Karsten Wade <kwade@redhat.com>

* New mailing list is live

 So far this is the only link from the website.

* Revert "Change link to point at new mailing list"

* Adding redirect, proper frontmatter

This should let us safely use this URL as canonical, knowing it's easy for a contributor to fix if the upstream URL changes, etc.